Extract interface for commitTsGetter | part 2
